Make tes3mobileActor.currentSpell usable + minor TES3 structure fixes.

+++ MWSE/TES3Alchemy.h
@@ -10,7 +10,7 @@
 
 namespace TES3 {
 	namespace AlchemyFlag {
-		typedef unsigned int value_type;
+		typedef unsigned short value_type;
 
 		enum Flag : value_type {
 			AutoCalc = 0x1
